test_requirements.txt
alabaster==0.7.12
appdirs==1.4.4
async-generator==1.10; python_version >= "3.5"
attrs==20.3.0; python_version >= "2.7" and python_version not in "3.0, 3.1, 3.2, 3.3"
babel==2.9.0; python_version >= "2.7" and python_version not in "3.0, 3.1, 3.2, 3.3"
black==20.8b1
bleach==3.3.0; python_version >= "2.7" and python_version not in "3.0, 3.1, 3.2, 3.3, 3.4"
build==0.1.0; python_version >= "2.7" and python_version not in "3.0, 3.1, 3.2, 3.3, 3.4"
certifi==2020.12.5
chardet==4.0.0; python_version >= "2.7" and python_version not in "3.0, 3.1, 3.2, 3.3, 3.4"
check-manifest==0.46
click==7.1.2; python_version >= "2.7" and python_version not in "3.0, 3.1, 3.2, 3.3, 3.4"
codespell==2.0.0
coverage==5.4; python_version >= "2.7" and python_version not in "3.0, 3.1, 3.2, 3.3, 3.4" and python_version < "4"
defusedxml==0.6.0; python_version >= "2.7" and python_version not in "3.0, 3.1, 3.2, 3.3, 3.4"
docutils==0.16; python_version >= "2.7" and python_version not in "3.0, 3.1, 3.2, 3.3, 3.4"
entrypoints==0.3; python_version >= "2.7"
idna==2.10; python_version >= "2.7" and python_version not in "3.0, 3.1, 3.2, 3.3"
imagesize==1.2.0; python_version >= "2.7" and python_version not in "3.0, 3.1, 3.2, 3.3"
importlib-metadata==3.4.0; python_version < "3.8"
iniconfig==1.1.1
ipython-genutils==0.2.0
jinja2==2.11.3
jsonschema==3.2.0
jupyter-client==6.1.11; python_version >= "3.5"
jupyter-core==4.7.1; python_version >= "3.6"
jupyterlab-pygments==0.1.2
markupsafe==1.1.1; python_version >= "2.7" and python_version not in "3.0, 3.1, 3.2, 3.3"
mistune==0.8.4
mypy==0.800
mypy-extensions==0.4.3
nbclient==0.5.1; python_version >= "3.6"
nbconvert==6.0.7; python_version >= "3.6"
nbformat==5.1.2; python_version >= "3.5"
nbsphinx==0.8.1
nest-asyncio==1.5.1; python_version >= "3.5"
numpydoc==1.1.0
packaging==20.9; python_version >= "2.7" and python_version not in "3.0, 3.1, 3.2, 3.3"
pandoc==1.0.2
pandocfilters==1.4.3
pathspec==0.8.1
pep517==0.9.1
pipfile-requirements==0.3.0
pluggy==0.13.1; python_version >= "2.7" and python_version not in "3.0, 3.1, 3.2, 3.3"
ply==3.11
py==1.10.0; python_version >= "2.7" and python_version not in "3.0, 3.1, 3.2, 3.3"
pycodestyle==2.6.0
pydocstyle==5.1.1
pygments==2.7.4; python_version >= "3.5"
pymcubes==0.1.2
pyparsing==2.4.7; python_version >= "2.6" and python_version not in "3.0, 3.1, 3.2, 3.3"
pyrsistent==0.17.3; python_version >= "3.5"
pytest==6.2.2
pytest-cov==2.11.1
python-dateutil==2.8.1; python_version >= "2.7" and python_version not in "3.0, 3.1, 3.2, 3.3"
pytz==2021.1
pyzmq==22.0.2; python_version >= "3.6"
regex==2020.11.13
requests==2.25.1; python_version >= "2.7" and python_version not in "3.0, 3.1, 3.2, 3.3, 3.4"
six==1.15.0; python_version >= "2.7" and python_version not in "3.0, 3.1, 3.2, 3.3"
snowballstemmer==2.1.0
sphinx==3.4.3
sphinx-autodoc-typehints==1.11.1
sphinx-bootstrap-theme==0.7.1
sphinx-copybutton==0.3.1
sphinx-gallery==0.8.2
sphinxcontrib-applehelp==1.0.2; python_version >= "3.5"
sphinxcontrib-devhelp==1.0.2; python_version >= "3.5"
sphinxcontrib-htmlhelp==1.0.3; python_version >= "3.5"
sphinxcontrib-jsmath==1.0.1; python_version >= "3.5"
sphinxcontrib-qthelp==1.0.3; python_version >= "3.5"
sphinxcontrib-restbuilder==0.2
sphinxcontrib-serializinghtml==1.1.4; python_version >= "3.5"
testpath==0.4.4
toml==0.10.2; python_version >= "2.6" and python_version not in "3.0, 3.1, 3.2, 3.3"
tornado==6.1; python_version >= "3.5"
traitlets==5.0.5; python_version >= "3.7"
typed-ast==1.4.2
typing-extensions==3.7.4.3; python_version < "3.8"
urllib3==1.26.3; python_version >= "2.7" and python_version not in "3.0, 3.1, 3.2, 3.3, 3.4" and python_version < "4"
webencodings==0.5.1
zipp==3.4.0; python_version < "3.8"
